requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
lxml==4.4.1
python-docx==0.8.10
requests==2.22.0
soupsieve==1.9.4
urllib3==1.25.6
